requirements.txt
django-colorfield
